// interface.ts
export interface SubscribeForPriceEvent {
    event: string;
    pair: string[];
    subscription: {
        name: string;
    };
}

export enum CoinType {
    ETH = 'eth',
    BTC = 'btc',
    LTC = 'ltc',
    DASH = 'dash',
    ZEC = 'zec',
}

export type CoinMap<T> = {
    [x in CoinType]: T;
};

export enum supportedPairs {
    ADA_CAD = 'ADA/CAD',
    ADA_ETH = 'ADA/ETH',
    ADA_EUR = 'ADA/EUR',
    ADA_USD = 'ADA/USD',
    ADA_XBT = 'ADA/XBT',
    ATOM_CAD = 'ATOM/CAD',
    ATOM_ETH = 'ATOM/ETH',
    ATOM_EUR = 'ATOM/EUR',
    ATOM_USD = 'ATOM/USD',
    ATOM_XBT = 'ATOM/XBT',
    BAT_ETH = 'BAT/ETH',
    BAT_EUR = 'BAT/EUR',
    BAT_USD = 'BAT/USD',
    BAT_XBT = 'BAT/XBT',
    BCH_EUR = 'BCH/EUR',
    BCH_USD = 'BCH/USD',
    BCH_XBT = 'BCH/XBT',
    BTC_CAD = 'BTC/CAD',
    BTC_EUR = 'BTC/EUR',
    BTC_GBP = 'BTC/GBP',
    BTC_JPY = 'BTC/JPY',
    BTC_USD = 'BTC/USD',
    DAI_EUR = 'DAI/EUR',
    DAI_USD = 'DAI/USD',
    DAI_USDT = 'DAI/USDT',
    DASH_EUR = 'DASH/EUR',
    DASH_USD = 'DASH/USD',
    DASH_XBT = 'DASH/XBT',
    DOGE_XBT = 'DOGE/XBT',
    EOS_ETH = 'EOS/ETH',
    EOS_EUR = 'EOS/EUR',
    EOS_USD = 'EOS/USD',
    EOS_XBT = 'EOS/XBT',
    ETC_ETH = 'ETC/ETH',
    ETC_EUR = 'ETC/EUR',
    ETC_USD = 'ETC/USD',
    ETC_XBT = 'ETC/XBT',
    ETH_CAD = 'ETH/CAD',
    ETH_DAI = 'ETH/DAI',
    ETH_EUR = 'ETH/EUR',
    ETH_GBP = 'ETH/GBP',
    ETH_JPY = 'ETH/JPY',
    ETH_USD = 'ETH/USD',
    ETH_XBT = 'ETH/XBT',
    GNO_ETH = 'GNO/ETH',
    GNO_EUR = 'GNO/EUR',
    GNO_USD = 'GNO/USD',
    GNO_XBT = 'GNO/XBT',
    ICX_ETH = 'ICX/ETH',
    ICX_EUR = 'ICX/EUR',
    ICX_USD = 'ICX/USD',
    ICX_XBT = 'ICX/XBT',
    LINK_ETH = 'LINK/ETH',
    LINK_EUR = 'LINK/EUR',
    LINK_USD = 'LINK/USD',
    LINK_XBT = 'LINK/XBT',
    LTC_EUR = 'LTC/EUR',
    LTC_USD = 'LTC/USD',
    LTC_XBT = 'LTC/XBT',
    MLN_ETH = 'MLN/ETH',
    MLN_XBT = 'MLN/XBT',
    QTUM_CAD = 'QTUM/CAD',
    QTUM_ETH = 'QTUM/ETH',
    QTUM_EUR = 'QTUM/EUR',
    QTUM_USD = 'QTUM/USD',
    QTUM_XBT = 'QTUM/XBT',
    REP_ETH = 'REP/ETH',
    REP_EUR = 'REP/EUR',
    REP_USD = 'REP/USD',
    REP_XBT = 'REP/XBT',
    SC_ETH = 'SC/ETH',
    SC_EUR = 'SC/EUR',
    SC_USD = 'SC/USD',
    SC_XBT = 'SC/XBT',
    STR_EUR = 'STR/EUR',
    STR_USD = 'STR/USD',
    STR_XBT = 'STR/XBT',
    USDT_USD = 'USDT/USD',
    WAVES_ETH = 'WAVES/ETH',
    WAVES_EUR = 'WAVES/EUR',
    WAVES_USD = 'WAVES/USD',
    WAVES_XBT = 'WAVES/XBT',
    XBT_CAD = 'XBT/CAD',
    XBT_EUR = 'XBT/EUR',
    XBT_GBP = 'XBT/GBP',
    XBT_JPY = 'XBT/JPY',
    XBT_USD = 'XBT/USD',
    XDG_XBT = 'XDG/XBT',
    XLM_EUR = 'XLM/EUR',
    XLM_USD = 'XLM/USD',
    XLM_XBT = 'XLM/XBT',
    XMR_EUR = 'XMR/EUR',
    XMR_USD = 'XMR/USD',
    XMR_XBT = 'XMR/XBT',
    XRP_CAD = 'XRP/CAD',
    XRP_EUR = 'XRP/EUR',
    XRP_JPY = 'XRP/JPY',
    XRP_USD = 'XRP/USD',
    XRP_XBT = 'XRP/XBT',
    XTZ_CAD = 'XTZ/CAD',
    XTZ_ETH = 'XTZ/ETH',
    XTZ_EUR = 'XTZ/EUR',
    XTZ_USD = 'XTZ/USD',
    XTZ_XBT = 'XTZ/XBT',
    ZEC_EUR = 'ZEC/EUR',
    ZEC_JPY = 'ZEC/JPY',
    ZEC_USD = 'ZEC/USD',
    ZEC_XBT = 'ZEC/XBT',
}